Create a dynamic way to provide component information
In some components, an action is offered that provides descriptive info about the component as well as links to the source code and the hig website. All this info has been gathered into a json file. The InfoData plugin has been created to serve the data to the QML code.

Differential Revision: https://phabricator.kde.org/D18823