Port Cards Grid, Cards List and Inline Messages extra information to json-based info sheet
Info sheet is fed by a single json file that contains extra info and links to source and hig. Cards Grid, Cards List and Inline Messages should use this to provide extra info. Also changed links that pointed to the kirigami repository, to kirigami-gallery which is more useful for the contributors.

Differential Revision: https://phabricator.kde.org/D18839